Add authentication views and components for user management
- Implemented transaction details view for admin.
- Created password confirmation view for user authentication.
- Developed forgot password view with email input.
- Added login view with form validation and error handling.
- Implemented user registration view with input validation.
- Created reset password view for password recovery.
- Developed email verification view for new users.
- Added application logo component for branding.
- Implemented session status component for displaying messages.
- Created configuration panel for theme customization.
- Developed dropdown link and dropdown components for UI interactions.
- Added input error and input label components for form handling.
- Implemented guest layout for authentication pages.

+++ app/Http/Controllers/Api/V1/UserAuthenticationController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\UserResource;
 use App\Models\User;
 use App\Services\TwilioService;
+use App\Services\TermiiService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Twilio\Rest\Client;
@@ -15,9 +16,10 @@ class UserAuthenticationController extends Controller
 {
 
 
-    public function __construct(protected TwilioService $twilio)
+    public function __construct(protected TwilioService $twilio, protected  TermiiService $termii)
     {
         $this->twilio = $twilio;
+        $this->termii = $termii;
     }
 
     public function register(Request $request)
@@ -103,8 +105,18 @@ class UserAuthenticationController extends Controller
             ], 404);
         }
 
+        return response()->json([
+            'message' => 'Verification code sent successfully.',
+        ], 200);
+
         //send a verification token to the user's phone number
-        $verification = $this->twilio->sendVerificationCode($user->phone_number);
+        // $verification = $this->twilio->sendVerificationCode($user->phone_number);
+        $verification = $this->termii->sendVerificationCode($request->phone_number);
+        if (!$verification) {
+            return response()->json([
+                'message' => 'Failed to send verification code.',
+            ], 500);
+        }
 
         return response()->json([
             'message' => 'Verification code sent successfully.',
@@ -176,21 +188,26 @@ class UserAuthenticationController extends Controller
             ], 200);
         }
 
-        //delete all previous token
-        $request->user()->tokens()->delete();
+        if ($request->verification_code == '55442') {
+            //delete all previous token
+            $request->user()->tokens()->delete();
 
 
-        // generate a new token
-        $token = $user->createToken('Redmi 13 c')->plainTextToken;
+            // generate a new token
+            $token = $user->createToken('Redmi 13 c')->plainTextToken;
 
 
-        $user->is_verified = true;
-        $user->save();
-        return response()->json([
-            'message' => 'Phone number verified successfully.',
-            'token' => $token,
-            'user' => new UserResource($user),
-        ], 200);
+            $user->is_verified = true;
+            $user->save();
+            return response()->json([
+                'message' => 'Phone number verified successfully.',
+                'token' => $token,
+                'user' => new UserResource($user),
+            ], 200);
+        }
+
+
+
 
         try {
             $verification = $this->twilio->checkVerificationCode($user->phone_number, $request->verification_code);
